Refactor DoublyLinkedList as many separated files
