+++ Gizmo/Classes/Common/DoublyLinkedList/DoublyLinkedList+Iterator.swift
@@ -1,0 +1,40 @@
+//
+//  DoublyLinkedList+Iterator.swift
+//  Gizmo
+//
+//  Created by Nayanda Haberty on 05/09/22.
+//
+
+import Foundation
+
+// MARK: DoublyLinkedListIterator
+
+public struct DoublyLinkedListIterator<Element>: IteratorProtocol {
+    
+    private var currentNode: DoublyLinkedList<Element>.Node?
+    
+    init(root: DoublyLinkedList<Element>.Node?) {
+        self.currentNode = root
+    }
+    
+    public mutating func next() -> Element? {
+        defer { currentNode = currentNode?.next }
+        return currentNode?.element
+    }
+}
+
+// MARK: DoublyLinkedListNodeIterator
+
+class DoublyLinkedListNodeIterator<Element>: SequenceWrapperIterator<DoublyLinkedList<Element>.Node> {
+    
+    private var currentNode: DoublyLinkedList<Element>.Node?
+    
+    init(root: DoublyLinkedList<Element>.Node?) {
+        self.currentNode = root
+    }
+    
+    override func next() -> DoublyLinkedList<Element>.Node? {
+        defer { currentNode = currentNode?.next }
+        return currentNode
+    }
+}
